+++ NEWS.md
@@ -1,8 +1,6 @@
 # cargo-mutants changelog
 
-## 0.0.3
-
-(Not yet released.)
+## unreleased
 
 * Skip functions or modules marked `#[test]`, `#[cfg(test)]` or
   `#[mutants::skip]`.
@@ -10,6 +8,8 @@
 * Early steps towards type-guided mutations: generate mutations of `true`
   and `false` for functions that return `bool`, and empty and arbitrary strings
   for functions returning `String`.
+
+* New `--list-mutants --json`.
 
 ## 0.0.2